Create 001-Minimum Jumps - cooding ninja
You are standing at (0,0) and want to reach '(X, Y)' in the minimum number of jumps.Before you start jumping, you choose an integer 'K'. In one jump you can jump from '(A, B)' to '(A + K, B)' or '(A, B + K)'. Your task is to find the minimum number of jumps to reach '(X, Y)' starting from (0, 0).

Example:
Lets 'X' = 4, and 'Y' = 2.
We can choose k = 2 and reach (4, 2) in 3 jumps.In the first jump, we go from (0, 0) to (0, 2). In the second jump, we go from (0, 2) to (2, 2). Hence, the answer is 3 jumps. In the third jump, we go from (2, 2) to (4, 2). Hence, the answer is 3 jumps.